fix segfault in malloc+strcpy, utils/chdir_keypath.c
git-svn-id: file:///Users/avp/tmp/svn.20121212/LHPC/aff/trunk@363 65d2813f-f11f-0410-b17b-8647a28db151

+++ utils/common.c
@@ -120,7 +120,7 @@ char *mk_tmp_filename( const char *mark, const char *fname )
 int split_farg( char *buf, int fargc, char ** fargv )
 {
     int part = 1;
-    for( size_t i = 1 ; i < fargc ; ++i )
+    for( int i = 1 ; i < fargc ; ++i )
         fargv[i] = NULL;
     fargv[0] = buf;
     int prev_bslash = 0,
